Fixed core lib schedule API signature

+++ org.lflang/src/lib/core/reactor.c
@@ -61,7 +61,7 @@ handle_t _lf_schedule_token(void* action, interval_t extra_delay, lf_token_t* to
  * Variant of schedule_token that creates a token to carry the specified value.
  * See reactor.h for documentation.
  */
-handle_t _lf_schedule_value(void* action, interval_t extra_delay, void* value, int length) {
+handle_t _lf_schedule_value(void* action, interval_t extra_delay, void* value, size_t length) {
     trigger_t* trigger = _lf_action_to_trigger(action);
     lf_token_t* token = create_token(trigger->element_size);
     token->value = value;
@@ -74,7 +74,7 @@ handle_t _lf_schedule_value(void* action, interval_t extra_delay, void* value, i
  * with a copy of the specified value.
  * See reactor.h for documentation.
  */
-handle_t _lf_schedule_copy(void* action, interval_t offset, void* value, int length) {
+handle_t _lf_schedule_copy(void* action, interval_t offset, void* value, size_t length) {
     trigger_t* trigger = _lf_action_to_trigger(action);
     if (value == NULL) {
         return schedule_token(action, offset, NULL);
